hugePack: Refactor from SocialMediaGallery to ImageGallery enhancing ImageGallery style and responsive add closeup view with popup to images install react modal
