+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "dev:expose": "vite --host 192.168.1.10",
+    "dev:expose": "vite --host 192.168.1.2",
     "build": "vite build",
     "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
@@ -20,6 +20,7 @@
     "react-bootstrap": "^2.8.0",
     "react-dom": "^18.2.0",
     "react-html-parser": "^2.0.2",
+    "react-modal": "^3.16.1",
     "react-router-dom": "^6.14.1"
   },
   "devDependencies": {
